BuildModel/requirements.txt
absl-py
astor
attrs
Babel
bleach
certifi
click
cycler
Flask
Flask-Babel
Flask-SQLAlchemy
Flask-Table
gast
grpcio
h5py
html5lib
itsdangerous
Jinja2
Keras
kiwisolver
Markdown
MarkupSafe
numpy
pandas
pluggy
protobuf
psycopg2
py
pyparsing
pytest
python-dateutil
pytz
PyYAML
scikit-learn
scipy
six
sklearn
Sphinx
sphinx-rtd-theme
sphinxcontrib-napoleon
SQLAlchemy
tensorboard
tensorflow
termcolor
virtualenv
Werkzeug
